+++ swarm/src/connection.rs
@@ -552,7 +552,7 @@ enum Shutdown {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::handler::KeepAliveConnectionHandler;
+    use crate::keep_alive;
     use futures::AsyncRead;
     use futures::AsyncWrite;
     use libp2p_core::upgrade::DeniedUpgrade;
@@ -572,7 +572,7 @@ mod tests {
                 StreamMuxerBox::new(DummyStreamMuxer {
                     counter: alive_substream_counter.clone(),
                 }),
-                KeepAliveConnectionHandler,
+                keep_alive::ConnectionHandler,
                 None,
                 max_negotiating_inbound_streams,
             );