+++ swarm/src/handler.rs
@@ -38,7 +38,6 @@
 //! >           the network as a whole, see the
 //! >           [`NetworkBehaviour`](crate::behaviour::NetworkBehaviour) trait.
 
-mod dummy;
 pub mod either;
 mod keep_alive;
 mod map_in;
@@ -54,8 +53,6 @@ use instant::Instant;
 use libp2p_core::{upgrade::UpgradeError, ConnectedPoint, Multiaddr, PeerId};
 use std::{cmp::Ordering, error, fmt, task::Context, task::Poll, time::Duration};
 
-pub use dummy::DummyConnectionHandler;
-pub use keep_alive::KeepAliveConnectionHandler;
 pub use map_in::MapInEvent;
 pub use map_out::MapOutEvent;
 pub use one_shot::{OneShotHandler, OneShotHandlerConfig};